Separate loungeMute from admin
