+++ src/components/admin/AdminLoungeMute.vue
@@ -117,13 +117,13 @@
 <script lang="ts">
 import Vue from "vue";
 import { Component } from "vue-property-decorator";
-import { LoungeMute, LoungeMuteResponse } from "@/store/admin/types";
+import { LoungeMute, LoungeMuteResponse } from "@/store/admin/loungeMute/types";
 import PlayerSearch from "@/components/common/PlayerSearch.vue";
 import { useOauthStore } from "@/store/oauth/store";
 import { setDate, setMonth, setYear } from "date-fns";
 
 @Component({ components: { PlayerSearch } })
-export default class AdmiLoungeMute extends Vue {
+export default class AdminLoungeMute extends Vue {
   public dateMenu = false;
   public dialog = false;
   public showConfirmation = false;
@@ -140,12 +140,12 @@ export default class AdmiLoungeMute extends Vue {
   ];
 
   get loungeMutes(): LoungeMuteResponse[] {
-    return this.$store.direct.state.admin.loungeMutedPlayers;
+    return this.$store.direct.state.loungeMute.loungeMutedPlayers;
   }
 
   async deleteItem(item: LoungeMute): Promise<void> {
     confirm("Are you sure you want to delete this item?") &&
-    await this.$store.direct.dispatch.admin.deleteLoungeMute(item.battleTag);
+    await this.$store.direct.dispatch.loungeMute.deleteLoungeMute(item.battleTag);
     this.loadMutes();
   }
 
@@ -167,7 +167,7 @@ export default class AdmiLoungeMute extends Vue {
       endDate: this.muteEndDateTime,
     } as LoungeMute;
 
-    await this.$store.direct.dispatch.admin.addLoungeMute(mute);
+    await this.$store.direct.dispatch.loungeMute.addLoungeMute(mute);
     this.loadMutes();
   }
 
@@ -185,7 +185,7 @@ export default class AdmiLoungeMute extends Vue {
   }
 
   public async loadMutes(): Promise<void> {
-    await this.$store.direct.dispatch.admin.loadLoungeMutes();
+    await this.$store.direct.dispatch.loungeMute.loadLoungeMutes();
   }
 
   async mounted(): Promise<void> {
